+++ CHANGES.md
@@ -3,6 +3,7 @@
   - https://github.com/koyeb/koyeb-cli/pull/250/files
 * Add the `--snapshot` flag to the volume creation command, to create a volume from a snapshot.
   - https://github.com/koyeb/koyeb-cli/pull/254/files
+  - https://github.com/koyeb/koyeb-cli/pull/255/files
 
 ## v5.1.0 (2024-09-26)